Change rc_override to be more symbolic (see below)
In the previous version of this API I used the 'tried and true'
rc_override terminology.  However I've changed rc_override to be
channel_override with a dictionary as the argument.
(This idea is from @rmackay9 - thanks!)

Strings will be defined per vehicle type ('pitch', 'yaw', 'roll' etc...)
and rather than setting channel 3 to 1400 (for instance), you will pass
in a dict with 'throttle':1200.  If you do not want to override particular
channels, you should not populate them in the dictionary.

It is worth noting by using a single dictionary it is guaranteed that all
multi-channel changes are updated atomically.

This change will be nice in two ways:

* we can hide (eventually we can deprecate) any notion of rc channel
numbers at all.
* vehicles can eventually define new 'channels' for overridden values.